:sparkles: feat(login): 删除用户时清理关联记录
在 RecordDao 中新增 deleteByUserId 方法，并在 User 控制器的 remove 方法中调用，确保删除用户时同步清理关联的记录数据。

+++ controller/User.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace nova\plugin\login\controller;
 
 use nova\framework\http\Response;
+use nova\plugin\login\db\Dao\RecordDao;
 use nova\plugin\login\db\Dao\RoleDao;
 use nova\plugin\login\db\Dao\UserDao;
 use nova\plugin\login\db\Model\UserModel;
@@ -104,7 +105,13 @@ class User extends BaseAPIController
             return Response::asJson(['code' => 404, 'msg' => '用户不存在']);
         }
 
+        if ($this->userModel->id === $user->id) {
+            return Response::asJson(['code' => 400, 'msg' => '不能删除自己']);
+        }
+
         UserDao::getInstance()->deleteById($id);
+
+        RecordDao::getInstance()->deleteByUserId($id);
 
         return Response::asJson(['code' => 200, 'msg' => '删除成功']);
     }
